update in mac

+++ install.sh
@@ -7,27 +7,43 @@ cd ~/.vim
 
 echo "create .vimrc"
 
+sysOS=`uname -s`
+
 mkdir ~/back
+function backup {
+  if [ -f ~/$i ]; then
+    mv ~/$i ~/back/${i}_`date +%Y%m%d_%s`
+  fi
+}
 
-if [ -f ~/.vimrc ]; then
-mv ~/.vimrc ~/back/.vimrc_`date +%Y%m%d_%s`
-fi
-ln -s ~/.vim/others/vimrc.local ~/.vimrc
+backup .vimrc
+backup .gvimrc
+backup .ideagvimrc
 
-if [ -f ~/.grvimrc ]; then
-mv ~/.gvimrc ~/.gvimrc_`date +%Y%m%d_%s`
+if [ $sysOS == "Darwin" ];then
+  echo "I'm MacOS"
+  ln -s ~/.vim/others/vimrc.local_mac ~/.vimrc
+  ln -s ~/.vim/others/vimrc.local_mac ~/.gvimrc
+  ln -s ~/.vim/others/vimrc.local_mac ~/.ideavimrc
+else
+  echo "Other OS: $sysOS"
+  ln -s ~/.vim/others/vimrc.local ~/.vimrc
+  ln -s ~/.vim/others/vimrc.local ~/.gvimrc
 fi
-ln -s ~/.vim/others/vimrc.local ~/.gvimrc
+ 
 
-if [ -f ~/.screenrc ]; then
-    mv ~/.screenrc ~/.screenrc_`date +%Y%m%d_%s`
-fi
+backup .screenrc
 ln -s ~/.vim/others/screenrc ~/.screenrc
 
-if [ -f ~/.bash_aliases ]; then
-    mv ~/.bash_aliases ~/.bash_aliases_`date +%Y%m%d_%s`
+backup .bash_alias
+if [ $sysOS == "Darwin" ];then
+  echo "I'm MacOS"
+  ln -s ~/.vim/others/bash_aliases_mac ~/.bash_aliases
+else
+  echo "Other OS: $sysOS"
+  ln -s ~/.vim/others/bash_aliases ~/.bash_aliases
 fi
-ln -s ~/.vim/others/bash_aliases ~/.bash_aliases
+ 
 
 if [ "`cat ~/.bashrc|grep -c bash_aliases`" != 0 ] ; then
 echo 'if [ -f ~/.bash_aliases ]; then
@@ -36,12 +52,14 @@ fi' >> ~/.bashrc
 fi
 
 
-mkdir ~/bin
-rm ~/bin/csg.sh
-ln -s ~/.vim/others/csg.sh ~/bin/csg.sh
+if [ $sysOS == "Linux" ];then
+  mkdir ~/bin
+  rm ~/bin/csg.sh
+  ln -s ~/.vim/others/csg.sh ~/bin/csg.sh
 
-rm ~/bin/csm.sh
-ln -s ~/.vim/others/csm.sh ~/bin/csm.sh
+  rm ~/bin/csm.sh
+  ln -s ~/.vim/others/csm.sh ~/bin/csm.sh
+fi
 
 git remote set-url origin git@github.com:lastagile/vim.git
